+++ lib/src/user_exception.dart
@@ -64,13 +64,6 @@ class UserException implements Exception {
 
   const UserException(this.msg, {this.cause, this.code});
 
-  /// Returns a deep copy of this exception, but stopping at, and not
-  /// including, the first [cause] which is not a UserException.
-  UserException pure() => UserException(
-        msg,
-        cause: (cause is UserException) ? (cause as UserException).pure() : null,
-      );
-
   /// Returns the first cause which, recursively, is NOT a UserException.
   /// If not found, returns null.
   Object hardCause() {
@@ -79,6 +72,13 @@ class UserException implements Exception {
     else
       return cause;
   }
+
+  /// Returns a deep copy of this exception, but stopping at, and not
+  /// including, the first [cause] which is not a UserException.
+  UserException withoutHardCause() => UserException(
+        msg,
+        cause: (cause is UserException) ? (cause as UserException).withoutHardCause() : null,
+      );
 
   String dialogTitle([Locale locale]) =>
       (cause is UserException || cause is String) ? _codeAsTextOrMsg(locale) : "";